Revert "feature/12-add-secure-keys-feature (#14)"
This reverts commit ecf33a44abdcca804dc50bfc6fce6106d0475027.
