+++ .gitignore
@@ -1,2 +1,2 @@
-.encrypted_keys
-.decrypted_keys
+watchtower-admin
+watchtower-operator
