+++ watchtower-operator-cli/ABIs/OperatorRegistry.abi
