+++ watchtower-operator-cli/bindings/AvsDirectory/AvsDirectory.go
