+++ watchtower-operator-cli/bindings/OperatorRegistry/OperatorRegistry.go
