+++ watchtower-operator-cli/bindings/WitnessHub/WitnessHub.go
