+++ watchtower-operator-cli/config/l1-operator-config.json.template
@@ -1,17 +1,15 @@
 {
   "watchtower_private_keys": [
-    "<WatchTower1-private-key-name>",
-    "<WatchTower2-private-key-name>",
-    "<WatchTower3-private-key-name>"
+    "<Watchtower Private Key 1>",
+    "<Watchtower Private Key 2>"
   ],
-  "operator_private_key": "<operator-private-key-name>",
+  "operator_private_key": "<Operator Private Key>",
   "operator_registry_address": "0xEf1a89841fd189ba28e780A977ca70eb1A5e985D",
   "witnesshub_address": "0xD25c2c5802198CB8541987b73A8db4c9BCaE5cC7",
   "avs_directory_address": "0x135dda560e946695d6f155dacafc6f1f25c1f5af",
   "eth_rpc_url": "<Mainnet RPC URL>",
   "chain_id": 1,
-  "gas_limit": "<gas-limit>",
-  "tx_receipt_timeout": "<timeout_in_seconds>",
-  "expiry": "<expiry_in_days>",
-  "use_encrypted_keys": false
+  "gas_limit": 500000,
+  "tx_receipt_timeout": 600,
+  "expiry_in_days": 1
 }
