+++ watchtower-operator-cli/config/l2-operator-config.json.template
@@ -1,0 +1,15 @@
+{
+  "watchtower_private_keys": [
+    "<WatchTower1-private-key>",
+    "<WatchTower2-private-key>",
+    "<WatchTower3-private-key>"
+  ],
+  "operator_private_key": "<operator-private-key>",
+  "operator_registry_address": "0xEf1a89841fd189ba28e780A977ca70eb1A5e985D",
+  "eth_rpc_url": "<rpc-url>",
+  "chain_id": "<chainID>",
+  "gas_limit": 500000,
+  "tx_receipt_timeout": 600,
+  "expiry_in_days": 1
+}
+
