+++ watchtower-operator-cli/go.mod
@@ -1,12 +1,11 @@
-module operator-cli
+module watchtower-operator
 
 go 1.22.1
 
 require (
 	github.com/ethereum/go-ethereum v1.13.14
-	github.com/howeyc/gopass v0.0.0-20210920133722-c8aef6fb66ef
-	github.com/urfave/cli/v2 v2.25.7
-	github.com/wagslane/go-password-validator v0.3.0
+	github.com/urfave/cli v1.22.14
+	gopkg.in/yaml.v2 v2.4.0
 )
 
 require (
@@ -32,13 +31,11 @@ require (
 	github.com/supranational/blst v0.3.11 // indirect
 	github.com/tklauser/go-sysconf v0.3.12 // indirect
 	github.com/tklauser/numcpus v0.6.1 // indirect
-	github.com/xrash/smetrics v0.0.0-20201216005158-039620a65673 // indirect
 	golang.org/x/crypto v0.17.0 // indirect
 	golang.org/x/exp v0.0.0-20231110203233-9a3e6036ecaa // indirect
 	golang.org/x/mod v0.14.0 // indirect
 	golang.org/x/sync v0.5.0 // indirect
 	golang.org/x/sys v0.16.0 // indirect
-	golang.org/x/term v0.15.0 // indirect
 	golang.org/x/tools v0.15.0 // indirect
 	rsc.io/tmplfunc v0.0.3 // indirect
 )
