+++ watchtower-operator-cli/release/version.go
@@ -1,0 +1,5 @@
+package release
+
+// var VERSION = strings.Replace(`
+// UNKNOWN
+// `,"\n","",-1)
