--- watchtower-operator/release/version.go
@@ -1,9 +1,0 @@
-package release
-
-import "strings"
-
-var VERSION = strings.Replace(`
-feat/25-give-option-to-the-user-to-use-encrypted-keys-or-not
--
-1713851371
-`,"\n","",-1)
